--- src/galax/dynamics/__init__.pyi
@@ -1,28 +1,0 @@
-# TODO: use star imports when
-# https://github.com/scientific-python/lazy_loader/issues/94 is resolved
-
-__all__ = [
-    # Modules
-    "integrate",
-    "mockstream",
-    # orbit, et al.
-    "AbstractOrbit",
-    "Orbit",
-    "InterpolatedOrbit",
-    # integrate
-    "evaluate_orbit",
-    # mockstream
-    "MockStreamArm",
-    "MockStream",
-    "MockStreamGenerator",
-    # mockstream.df
-    "AbstractStreamDF",
-    "FardalStreamDF",
-]
-
-from ._dynamics import integrate, mockstream
-from ._dynamics.base import AbstractOrbit
-from ._dynamics.integrate._funcs import evaluate_orbit
-from ._dynamics.mockstream import MockStream, MockStreamArm, MockStreamGenerator
-from ._dynamics.mockstream.df import AbstractStreamDF, FardalStreamDF
-from ._dynamics.orbit import InterpolatedOrbit, Orbit
