+++ src/galax/dynamics/_dynamics/integrate/__init__.py
@@ -1,16 +1,16 @@
 """galax: Galactic Dynamix in Jax."""
 
-from . import _api, _base, _builtin, _funcs
-from ._api import *
-from ._base import *
-from ._builtin import *
-from ._funcs import *
+from . import api, base, builtin, funcs
+from .api import *
+from .base import *
+from .builtin import *
+from .funcs import *
 
 __all__: list[str] = []
-__all__ += _api.__all__
-__all__ += _base.__all__
-__all__ += _builtin.__all__
-__all__ += _funcs.__all__
+__all__ += api.__all__
+__all__ += base.__all__
+__all__ += builtin.__all__
+__all__ += funcs.__all__
 
 # Cleanup
-del _api, _base, _builtin, _funcs
+del api, base, builtin, funcs
