+++ src/galax/dynamics/_dynamics/mockstream/df/__init__.py
@@ -1,14 +1,14 @@
 """galax: Galactic Dynamix in Jax."""
 
-from . import _base, _fardal15, _progenitor
-from ._base import *
-from ._fardal15 import *
-from ._progenitor import *
+from . import base, fardal15, progenitor
+from .base import *
+from .fardal15 import *
+from .progenitor import *
 
 __all__: list[str] = []
-__all__ += _base.__all__
-__all__ += _progenitor.__all__
-__all__ += _fardal15.__all__
+__all__ += base.__all__
+__all__ += progenitor.__all__
+__all__ += fardal15.__all__
 
 # Cleanup
-del _base, _fardal15, _progenitor
+del base, fardal15, progenitor
