+++ src/galax/dynamics/_dynamics/mockstream/df/base.py
@@ -17,7 +17,7 @@ from unxt import Quantity
 import galax.coordinates as gc
 import galax.potential as gp
 import galax.typing as gt
-from ._progenitor import ConstantMassProtenitor, ProgenitorMassCallable
+from .progenitor import ConstantMassProtenitor, ProgenitorMassCallable
 from galax.dynamics._dynamics.mockstream.core import MockStreamArm
 from galax.dynamics._dynamics.orbit import Orbit
