+++ src/galax/dynamics/_dynamics/mockstream/df/progenitor.py
