+++ src/galax/dynamics/integrate.py
@@ -1,0 +1,10 @@
+""":mod:`galax.dynamics.integrate`."""
+
+from ._dynamics import integrate
+from ._dynamics.integrate import *  # noqa: F403
+
+__all__: list[str] = []
+__all__ += integrate.__all__
+
+# Cleanup
+del integrate
