+++ src/galax/dynamics/mockstream.py
@@ -1,0 +1,10 @@
+""":mod:`galax.dynamics.mockstream`."""
+
+from ._dynamics import mockstream
+from ._dynamics.mockstream import *  # noqa: F403
+
+__all__: list[str] = []
+__all__ += mockstream.__all__
+
+# Cleanup
+del mockstream
