--- src/galax/potential/__init__.pyi
@@ -1,120 +1,0 @@
-# TODO: use star imports when
-# https://github.com/scientific-python/lazy_loader/issues/94 is resolved
-
-__all__ = [
-    # Modules
-    "io",
-    # base
-    "AbstractPotentialBase",
-    # core
-    "AbstractPotential",
-    # composite
-    "AbstractCompositePotential",
-    "CompositePotential",
-    # builtin
-    "BurkertPotential",
-    "HernquistPotential",
-    "IsochronePotential",
-    "JaffePotential",
-    "KeplerPotential",
-    "KuzminPotential",
-    "MiyamotoNagaiPotential",
-    "NullPotential",
-    "PlummerPotential",
-    "PowerLawCutoffPotential",
-    "SatohPotential",
-    "StoneOstriker15Potential",
-    "TriaxialHernquistPotential",
-    # bars
-    "BarPotential",
-    "LongMuraliBarPotential",
-    # logarithmic
-    "LogarithmicPotential",
-    "LMJ09LogarithmicPotential",
-    # nfw
-    "NFWPotential",
-    "LeeSutoTriaxialNFWPotential",
-    "TriaxialNFWPotential",
-    "Vogelsberger08TriaxialNFWPotential",
-    # special
-    "BovyMWPotential2014",
-    "LM10Potential",
-    "MilkyWayPotential",
-    # frame
-    "PotentialFrame",
-    # funcs
-    "potential",
-    "gradient",
-    "laplacian",
-    "density",
-    "hessian",
-    "acceleration",
-    "tidal_tensor",
-    # param
-    "AbstractParametersAttribute",
-    "ParametersAttribute",
-    "CompositeParametersAttribute",
-    "ParameterCallable",
-    "AbstractParameter",
-    "ConstantParameter",
-    "LinearParameter",
-    "UserParameter",
-    "ParameterField",
-]
-
-from ._potential import io
-from ._potential.base import AbstractPotentialBase
-from ._potential.builtin.bars import BarPotential, LongMuraliBarPotential
-from ._potential.builtin.builtin import (
-    BurkertPotential,
-    HernquistPotential,
-    IsochronePotential,
-    JaffePotential,
-    KeplerPotential,
-    KuzminPotential,
-    MiyamotoNagaiPotential,
-    NullPotential,
-    PlummerPotential,
-    PowerLawCutoffPotential,
-    SatohPotential,
-    StoneOstriker15Potential,
-    TriaxialHernquistPotential,
-)
-from ._potential.builtin.logarithmic import (
-    LMJ09LogarithmicPotential,
-    LogarithmicPotential,
-)
-from ._potential.builtin.nfw import (
-    LeeSutoTriaxialNFWPotential,
-    NFWPotential,
-    TriaxialNFWPotential,
-    Vogelsberger08TriaxialNFWPotential,
-)
-from ._potential.builtin.special import (
-    BovyMWPotential2014,
-    LM10Potential,
-    MilkyWayPotential,
-)
-from ._potential.composite import AbstractCompositePotential, CompositePotential
-from ._potential.core import AbstractPotential
-from ._potential.frame import PotentialFrame
-from ._potential.funcs import (
-    acceleration,
-    density,
-    gradient,
-    hessian,
-    laplacian,
-    potential,
-    tidal_tensor,
-)
-from ._potential.param import (
-    AbstractParameter,
-    AbstractParametersAttribute,
-    CompositeParametersAttribute,
-    ConstantParameter,
-    LinearParameter,
-    ParameterCallable,
-    ParameterField,
-    ParametersAttribute,
-    UserParameter,
-)
