+++ src/galax/potential/_potential/__init__.py
@@ -1,119 +1,3 @@
 """``galax`` Potentials."""
 
-__all__ = [
-    # Modules
-    "io",
-    # base
-    "AbstractPotentialBase",
-    # core
-    "AbstractPotential",
-    # composite
-    "AbstractCompositePotential",
-    "CompositePotential",
-    # builtin
-    "BurkertPotential",
-    "HernquistPotential",
-    "IsochronePotential",
-    "JaffePotential",
-    "KeplerPotential",
-    "KuzminPotential",
-    "MiyamotoNagaiPotential",
-    "NullPotential",
-    "PlummerPotential",
-    "PowerLawCutoffPotential",
-    "SatohPotential",
-    "StoneOstriker15Potential",
-    "TriaxialHernquistPotential",
-    # bars
-    "BarPotential",
-    "LongMuraliBarPotential",
-    # logarithmic
-    "LogarithmicPotential",
-    "LMJ09LogarithmicPotential",
-    # nfw
-    "NFWPotential",
-    "LeeSutoTriaxialNFWPotential",
-    "TriaxialNFWPotential",
-    "Vogelsberger08TriaxialNFWPotential",
-    # special
-    "BovyMWPotential2014",
-    "LM10Potential",
-    "MilkyWayPotential",
-    # frame
-    "PotentialFrame",
-    # funcs
-    "potential",
-    "gradient",
-    "laplacian",
-    "density",
-    "hessian",
-    "acceleration",
-    "tidal_tensor",
-    # param
-    "AbstractParametersAttribute",
-    "ParametersAttribute",
-    "CompositeParametersAttribute",
-    "ParameterCallable",
-    "AbstractParameter",
-    "ConstantParameter",
-    "LinearParameter",
-    "UserParameter",
-    "ParameterField",
-]
-
-from . import io
-from .base import AbstractPotentialBase
-from .builtin.bars import BarPotential, LongMuraliBarPotential
-from .builtin.builtin import (
-    BurkertPotential,
-    HernquistPotential,
-    IsochronePotential,
-    JaffePotential,
-    KeplerPotential,
-    KuzminPotential,
-    MiyamotoNagaiPotential,
-    NullPotential,
-    PlummerPotential,
-    PowerLawCutoffPotential,
-    SatohPotential,
-    StoneOstriker15Potential,
-    TriaxialHernquistPotential,
-)
-from .builtin.logarithmic import (
-    LMJ09LogarithmicPotential,
-    LogarithmicPotential,
-)
-from .builtin.nfw import (
-    LeeSutoTriaxialNFWPotential,
-    NFWPotential,
-    TriaxialNFWPotential,
-    Vogelsberger08TriaxialNFWPotential,
-)
-from .builtin.special import (
-    BovyMWPotential2014,
-    LM10Potential,
-    MilkyWayPotential,
-)
-from .composite import AbstractCompositePotential, CompositePotential
-from .core import AbstractPotential
-from .frame import PotentialFrame
-from .funcs import (
-    acceleration,
-    density,
-    gradient,
-    hessian,
-    laplacian,
-    potential,
-    tidal_tensor,
-)
-from .param import (
-    AbstractParameter,
-    AbstractParametersAttribute,
-    CompositeParametersAttribute,
-    ConstantParameter,
-    LinearParameter,
-    ParameterCallable,
-    ParameterField,
-    ParametersAttribute,
-    UserParameter,
-)
+__all__: list[str] = []
