--- src/galax/potential/_potential/param/__init__.py
@@ -1,11 +1,0 @@
-from . import attr, core, field, utils
-from .attr import *
-from .core import *
-from .field import *
-from .utils import *
-
-__all__: list[str] = []
-__all__ += attr.__all__
-__all__ += core.__all__
-__all__ += field.__all__
-__all__ += utils.__all__
