+++ src/galax/potential/_potential/params/__init__.py
@@ -1,0 +1,3 @@
+""":mod:`galax.potential.params`."""
+
+__all__: list[str] = []
