+++ src/galax/potential/_potential/params/attr.py
