+++ src/galax/potential/io.py
@@ -1,0 +1,16 @@
+""":mod:`galax.potential.params`."""
+
+__all__ = [
+    "convert_potential",
+    "AbstractInteroperableLibrary",
+    "GalaxLibrary",
+    "GalaLibrary",
+]
+
+
+from ._potential.io import (
+    AbstractInteroperableLibrary,
+    GalaLibrary,
+    GalaxLibrary,
+    convert_potential,
+)
