+++ src/galax/potential/params.py
@@ -1,0 +1,28 @@
+""":mod:`galax.potential.params`."""
+
+__all__ = [
+    "AbstractParametersAttribute",
+    "ParametersAttribute",
+    "CompositeParametersAttribute",
+    "ParameterCallable",
+    "AbstractParameter",
+    "ConstantParameter",
+    "LinearParameter",
+    "UserParameter",
+    "ParameterField",
+]
+
+
+from ._potential.params.attr import (
+    AbstractParametersAttribute,
+    CompositeParametersAttribute,
+    ParametersAttribute,
+)
+from ._potential.params.core import (
+    AbstractParameter,
+    ConstantParameter,
+    LinearParameter,
+    ParameterCallable,
+    UserParameter,
+)
+from ._potential.params.field import ParameterField
