+++ tests/smoke/dynamics/integrate/test_package.py
@@ -1,11 +1,11 @@
 """Test the :mod:`galax.integrate` module."""
 
 from galax.dynamics import integrate
-from galax.dynamics._dynamics.integrate import _api, _base, _builtin, _funcs
+from galax.dynamics._dynamics.integrate import api, base, builtin, funcs
 
 
 def test_all() -> None:
     """Test the API."""
     assert set(integrate.__all__) == set(
-        _api.__all__ + _base.__all__ + _builtin.__all__ + _funcs.__all__
+        api.__all__ + base.__all__ + builtin.__all__ + funcs.__all__
     )
